Share the streamed-upload helper between upload routes (#53)
World import owned the only chunked-with-a-cap upload reader in the tree.
The .mrpack import needs exactly the same thing, and a guard whose whole
job is to stop an oversized body filling the disk is the wrong code to
copy: two copies drift, and the copy that drifts is the one nobody is
looking at.

Moves it to backend/utils/upload.py along with the env-var cap reader
that was inlined next to it. world_import re-exports both names, so its
route, its callers and its tests keep working against the module they
already know.
